+++ controllers/content/ContentController.php
@@ -80,7 +80,4 @@ class ContentController extends BaseController
         }
         return $this->returnError(500, 'Internal error while delete content!');
     }
-
-
-
 }