+++ docs/CHANGELOG.md
@@ -7,7 +7,8 @@ Changelog
 - Fix #60: Cleanup code for prepare params
 - Enh #86: Docs for polls module
 - Fix #67: Update docs for container(space/user) tags
-- Enh #74: Response space container ID 
+- Enh #74: Response space container ID
+- Enh #75: Attach files to content
 
 
 0.3.1  (Unreleased)